Move session and fargment to lib
